Refactor menu handlers: introduce MainMenuHandler base class to unify shared menu logic
