+++ src/main/kotlin/ui/controller/CreateProjectUiController.kt
@@ -1,0 +1,7 @@
+package ui.controller
+
+class CreateProjectUiController: BaseUIController {
+    override fun execute() {
+        println("dummy project --------------")
+    }
+}
